wip - copycat implementation of MediachainClient
- needs testing
- pretty sure the latestBlockIndex logic could lead to
  gaps in the chain... need to sort that out.
