+++ client/src/main/scala/io/mediachain/client/MediachainClient.scala
@@ -1,15 +1,10 @@
 package io.mediachain.client
 
-import cats.data.{Xor, XorT}
+import cats.data.XorT
 import io.mediachain.protocol.Datastore._
-import io.mediachain.protocol.Transactor.{JournalError, JournalListener}
-import io.mediachain.transactor.Copycat.{ClientState, ClientStateListener}
-
-import scala.concurrent.{ExecutionContext, Future}
-import io.mediachain.protocol.Datastore._
+import io.mediachain.protocol.Transactor.JournalError
 
 import scala.concurrent.Future
-import scala.util.{Failure, Success}
 
 
 sealed trait MediachainClientEvent
@@ -53,6 +48,7 @@ trait MediachainClient {
   /**
     * Add an event listener to be informed of new and updated
     * mediachain records.
+    *
     * @param listener an event listener
     */
   def addListener(listener: ClientEventListener): Unit
@@ -60,15 +56,17 @@ trait MediachainClient {
 
   /**
     * Add a new CanonicalRecord to the system
+    *
     * @param canonicalRecord a new Entity or Artefact to add to the system
-    * @return an Xor-wrapped future that will complete with either a ClientError or a
-    *         Reference to the new record
+    * @return an Xor-wrapped future that will complete with either a ClientError
+    *         or a Reference to the new record
     */
   def addCanonical(canonicalRecord: CanonicalRecord): XorT[Future, ClientError, Reference]
 
 
   /**
     * Update an existing CanonicalRecord
+    *
     * @param canonicalReference a Reference to the canonical to update
     * @param chainCell a ChainCell containing the new metadata for the record.
     * @return an Xor-wrapped future that will complete with either a ClientError
@@ -78,108 +76,5 @@ trait MediachainClient {
   : XorT[Future, ClientError, Reference]
 }
 
-
-
-class MediachainCopycatClient(datastore: Datastore)
-  (implicit executionContext: ExecutionContext = ExecutionContext.global)
-  extends MediachainClient with JournalListener with ClientStateListener
-{
-  import io.mediachain.transactor.Copycat
-
-  def allCanonicalReferences = canonicalRefs
-
-  def chainForCanonical(ref: Reference): Future[Option[Reference]] = {
-    // TODO: handle disconnected cluster state
-    cluster.lookup(ref)
-  }
-
-  def addListener(listener: ClientEventListener): Unit = {
-    listeners += listener
-  }
-
-
-
-  /**
-    * Add a new CanonicalRecord to the system
-    *
-    * @param canonicalRecord a new Entity or Artefact to add to the system
-    * @return an Xor-wrapped future that will complete with either a ClientError or a
-    *         Reference to the new record
-    */
-  def addCanonical(canonicalRecord: CanonicalRecord)
-  : XorT[Future, ClientError, Reference] = ???
-
-  /**
-    * Update an existing CanonicalRecord
-    *
-    * @param canonicalReference a Reference to the canonical to update
-    * @param chainCell          a ChainCell containing the new metadata for the record.
-    * @return an Xor-wrapped future that will complete with either a ClientError
-    *         or a Reference to the new head of the record's chain
-    */
-  def updateCanonical(canonicalReference: Reference, chainCell: ChainCell): XorT[Future, ClientError, Reference] = ???
-
-
-var listeners: Set[ClientEventListener] = Set()
-  var canonicalRefs: Set[Reference] = Set()
-  var clusterClientState: ClientState = ClientState.Disconnected
-
-  val cluster = Copycat.Client.build()
-  cluster.listen(this)
-  cluster.addStateListener(this)
-
-  def connect(address: String) = {
-    cluster.connect(address)
-  }
-
-  def close() = cluster.close()
-
-  def onStateChange(state: ClientState): Unit = {
-    import ClientState._
-    clusterClientState = state
-
-    state match {
-      case Connected => catchupJournal()
-      case _ => ()
-    }
-  }
-
-  def catchupJournal(): Unit = {
-    cluster.currentBlock.map { block =>
-      block.entries.foreach(handleJournalEntry)
-      block.chain.foreach(backtrackJournal)
-    }
-  }
-
-  def backtrackJournal(blockRef: Reference): Unit = {
-    val block = datastore.getAs[JournalBlock](blockRef)
-      .getOrElse(
-        // TODO: better failure handling
-        throw new RuntimeException(s"Unable to fetch block with ref: $blockRef")
-      )
-    handleNewBlock(block)
-  }
-
-  private var latestBlockIndex: BigInt = -1
-  private def handleNewBlock(block: JournalBlock): Unit = {
-    block.entries.foreach(handleJournalEntry)
-
-    if (block.index > latestBlockIndex) {
-      block.chain.foreach(backtrackJournal)
-      latestBlockIndex = block.index
-    }
-  }
-
-  private def handleJournalEntry(entry: JournalEntry): Unit = {
-    canonicalRefs += entry.ref
-  }
-
-  override def onJournalCommit(entry: JournalEntry): Unit =
-    handleJournalEntry(entry)
-
-  override def onJournalBlock(ref: Reference): Unit = {
-    datastore.getAs[JournalBlock](ref).foreach(handleNewBlock)
-  }
-}
